# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shift Scheduler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /surveys:
    get:
      operationId: listSurveys
      tags: [surveys]
      summary: List all surveys
      responses:
        "200":
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
    post:
      operationId: createSurvey
      tags: [surveys]
      summary: Create a new survey for a given month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyBody"
      responses:
        "201":
          description: Survey created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /surveys/{id}:
    get:
      operationId: getSurvey
      tags: [surveys]
      summary: Get a survey with all its shifts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSurvey
      tags: [surveys]
      summary: Update a survey (e.g. close it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurveyBody"
      responses:
        "200":
          description: Survey updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /surveys/{id}/responses:
    get:
      operationId: getSurveyResponses
      tags: [responses]
      summary: Get all responses for a survey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespondentWithResponses"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /surveys/{id}/stats:
    get:
      operationId: getSurveyStats
      tags: [surveys]
      summary: Get availability statistics for a survey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /respondents:
    get:
      operationId: listRespondents
      tags: [respondents]
      summary: List all respondents
      responses:
        "200":
          description: List of respondents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Respondent"
    post:
      operationId: createRespondent
      tags: [respondents]
      summary: Create or update a respondent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRespondentBody"
      responses:
        "201":
          description: Respondent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respondent"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /respondents/{id}:
    patch:
      operationId: updateRespondent
      tags: [respondents]
      summary: Update a respondent's category or name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRespondentBody"
      responses:
        "200":
          description: Respondent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respondent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRespondent
      tags: [respondents]
      summary: Delete a respondent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Survey:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
          enum: [open, closed]
        token:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - month
        - year
        - status
        - token
        - createdAt
        - updatedAt
    Shift:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        date:
          type: string
          format: date
        dayType:
          type: string
          enum: [weekday, weekend]
        startTime:
          type: string
        endTime:
          type: string
        durationHours:
          type: number
        label:
          type: string
      required:
        - id
        - surveyId
        - date
        - dayType
        - startTime
        - endTime
        - durationHours
        - label
    SurveyDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
          enum: [open, closed]
        token:
          type: string
        shifts:
          type: array
          items:
            $ref: "#/components/schemas/Shift"
        responseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - month
        - year
        - status
        - token
        - shifts
        - responseCount
        - createdAt
        - updatedAt
    Respondent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        category:
          type: string
          enum: [AFP, General]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - createdAt
    RespondentWithResponses:
      type: object
      properties:
        respondentId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [AFP, General]
        selectedShiftIds:
          type: array
          items:
            type: integer
        totalAvailableHours:
          type: number
      required:
        - respondentId
        - name
        - category
        - selectedShiftIds
        - totalAvailableHours
...
    CreateSurveyBody:
      type: object
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        title:
          type: ["string", "null"]
      required:
        - month
        - year
    UpdateSurveyBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, closed]
        title:
          type: ["string", "null"]
    CreateRespondentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        category:
          type: string
          enum: [AFP, General]
      required:
        - name
        - category
    UpdateRespondentBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
          enum: [AFP, General, null]
...
    ShiftTypeStat:
      type: object
      properties:
        shiftLabel:
          type: string
        dayType:
          type: string
          enum: [weekday, weekend]
        totalSelections:
          type: integer
        selectionRate:
          type: number
      required:
        - shiftLabel
        - dayType
        - totalSelections
        - selectionRate
    RespondentStat:
      type: object
      properties:
        respondentId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [AFP, General]
        totalAvailableHours:
          type: number
        shiftsSelected:
          type: integer
        weekdayShifts:
          type: integer
        weekendShifts:
          type: integer
      required:
        - respondentId
        - name
        - category
        - totalAvailableHours
        - shiftsSelected
        - weekdayShifts
        - weekendShifts
    SurveyStats:
      type: object
      properties:
        totalRespondents:
          type: integer
        averageAvailableHours:
          type: number
        stdDevAvailableHours:
          type: number
        shiftTypeStats:
          type: array
          items:
            $ref: "#/components/schemas/ShiftTypeStat"
        respondentStats:
          type: array
          items:
            $ref: "#/components/schemas/RespondentStat"
      required:
        - totalRespondents
        - averageAvailableHours
        - stdDevAvailableHours
        - shiftTypeStats
        - respondentStats
...