# lib/api-zod/src/generated/types/respondent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shift Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondentCategory } from "./respondentCategory";

export interface Respondent {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  category: RespondentCategory;
  createdAt: Date;
}