# lib/api-zod/src/generated/types/respondentStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shift Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondentStatCategory } from "./respondentStatCategory";

export interface RespondentStat {
  respondentId: number;
  name: string;
  category: RespondentStatCategory;
  totalAvailableHours: number;
  shiftsSelected: number;
  weekdayShifts: number;
  weekendShifts: number;
}