# lib/api-zod/src/generated/types/shiftTypeStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shift Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftTypeStatDayType } from "./shiftTypeStatDayType";

export interface ShiftTypeStat {
  shiftLabel: string;
  dayType: ShiftTypeStatDayType;
  totalSelections: number;
  selectionRate: number;
}