# lib/api-zod/src/generated/types/surveyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shift Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondentStat } from "./respondentStat";
import type { ShiftTypeStat } from "./shiftTypeStat";

export interface SurveyStats {
  totalRespondents: number;
  averageAvailableHours: number;
  stdDevAvailableHours: number;
  shiftTypeStats: ShiftTypeStat[];
  respondentStats: RespondentStat[];
}