# lib/api-zod/src/generated/types/updateRespondentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shift Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateRespondentBodyCategory } from "./updateRespondentBodyCategory";

export interface UpdateRespondentBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  category?: UpdateRespondentBodyCategory;
}